backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
pypdf>=4.2.0
pdfminer.six>=20221105
sentence-transformers>=3.0.0
chromadb>=0.5.3
httpx>=0.27.0
numpy>=1.26.0
tenacity>=8.3.0

# Training dependencies (LoRA fine-tuning on GPU/MPS)
transformers>=4.44.0
peft>=0.12.0
trl>=0.10.0
datasets>=2.21.0
accelerate>=0.34.0
scipy>=1.13.0
